splitting reliable batch from regular batches
