+++ lib/sidekiq/grouping/batch.rb
@@ -41,8 +41,6 @@ module Sidekiq
       end
 
       def flush
-        return reliable_flush if reliable?
-
         chunk = pluck
         return unless chunk
 
@@ -54,29 +52,6 @@ module Sidekiq
           )
         end
         set_current_time_as_last
-      end
-
-      def reliable_flush
-        pending_name, chunk = reliable_pluck
-        return unless chunk
-
-        chunk.each_slice(chunk_size) do |subchunk|
-          Sidekiq::Client.push(
-            'class' => @worker_class,
-            'queue' => @queue,
-            'args' => [true, subchunk]
-          )
-        end
-        @redis.remove_from_pending(@name, pending_name)
-        set_current_time_as_last
-      end
-
-      def reliable_pluck
-        if @redis.lock(@name)
-          pending_name, items = @redis.reliable_pluck(@name, pluck_size)
-          items = items.map { |value| JSON.parse(value) }
-          [pending_name, items]
-        end
       end
 
       def worker_class_constant
@@ -109,10 +84,6 @@ module Sidekiq
         @redis.delete(@name)
       end
 
-      def requeue_expired
-        @redis.requeue_expired(@name, reliable_ttl)
-      end
-
       private
 
       def could_flush_on_overflow?
@@ -137,14 +108,6 @@ module Sidekiq
         worker_class_options['batch_unique'] == true
       end
 
-      def reliable?
-        worker_class_options['batch_reliable'] == true || Sidekiq::Grouping::Config.reliable == true
-      end
-
-      def reliable_ttl
-        worker_class_options['batch_reliable_ttl'] || 3600
-      end
-
       def set_current_time_as_last
         @redis.set_last_execution_time(@name, Time.now)
       end
@@ -154,8 +117,21 @@ module Sidekiq
           redis = Sidekiq::Grouping::Redis.new
 
           redis.batches.map do |name|
-            new(*extract_worker_klass_and_queue(name))
-          end
+            if Sidekiq::Grouping::Config.reliable
+              begin
+                klass, queue = extract_worker_klass_and_queue(name)
+                if klass.constantize.get_sidekiq_options['batch_reliable'] == true
+                  Sidekiq::Grouping::ReliableBatch.new(klass, queue)
+                else
+                  new(klass, queue)
+                end
+              rescue NameError
+                nil
+              end
+            else
+              new(*extract_worker_klass_and_queue(name))
+            end
+          end.compact
         end
 
         def extract_worker_klass_and_queue(name)
